% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcModule.R
\name{QcModule}
\alias{QcModule}
\title{Create an object of class QcModule}
\usage{
QcModule(data, tool = NULL, version = NULL, ...)
}
\arguments{
\item{data}{A data.frame containing data from a given bioinformatics tool}

\item{tool}{optional. The name of the tool used, e.g. "samtools stats"}

\item{version}{optional. The tool version(s) used as a character vector}

\item{...}{Not used}
}
\value{
An object of class QcModule
}
\description{
Create an object of class QcModule
}
\details{
This will coerce a data.frame or matrix into the correct structure to be a
QcModule. QcModule objects are data.frames set as S4 objects with optional
attributes 'tool' and 'version'
}
\examples{
## Create a simple object
mod <- QcModule(pressure, tool = "example")
mod
attributes(mod)


}
